64-bit prep part 4

+++ AudioScrobbler/AudioScrobblerClient.h
@@ -24,7 +24,7 @@
 
 @interface AudioScrobblerClient : NSObject
 {
-	int			_socket;
+	NSInteger			_socket;
 	BOOL		_doPortStepping;
 	in_port_t	_port;
 }
